(feat): add thread env.
