+++ examples/qwen2.5-0.5B-agentic_ds/agent_val_frozen_lake.yaml
@@ -32,7 +32,7 @@ system_envs:
 
 track_with: tensorboard
 tracker_kwargs:
-  log_dir: /data/oss_bucket_0/yali/llm/tensorboard/roll_exp/agentic_sokoban
+  log_dir: /data/oss_bucket_0/yali/llm/tensorboard/roll_exp/agentic_frozen_lake
 
 num_gpus_per_node: 8
 
@@ -42,7 +42,7 @@ logging_steps: 1
 eval_steps: 10
 resume_from_checkpoint: false
 
-rollout_batch_size: 512
+rollout_batch_size: 1024
 val_batch_size: 1024
 sequence_length: 8192
 
@@ -69,13 +69,20 @@ actor_train:
     learning_rate: 1.0e-6
     weight_decay: 0
     per_device_train_batch_size: 2
-    gradient_accumulation_steps: 32
+    gradient_accumulation_steps: 64
     warmup_steps: 10
   data_args:
     template: qwen2_5
   strategy_args:
-    strategy_name: deepspeed_train
-    strategy_config: ${deepspeed_zero2}
+#    strategy_name: deepspeed_train
+#    strategy_config: ${deepspeed_zero3}
+    strategy_name: megatron_train
+    strategy_config:
+      tensor_model_parallel_size: 1
+      pipeline_model_parallel_size: 1
+      expert_model_parallel_size: 1
+      use_distributed_optimizer: true
+      recompute_granularity: full
   device_mapping: list(range(0,8))
   infer_batch_size: 2
 
@@ -85,7 +92,7 @@ actor_infer:
     disable_gradient_checkpointing: true
     dtype: bf16
   generating_args:
-    max_new_tokens: 32 # single-turn response length
+    max_new_tokens: 128 # single-turn response length
     top_p: 0.99
     top_k: 100
     num_beams: 1
@@ -120,7 +127,7 @@ enable_response_mask: True
 action_sep: "||"
 use_turn_scores: False # important to GAE when applying token-level rewards to token-level advantages. If False, will take the sum of scores as the reward for the last turn.
 enable_think: False # False -> no think RL
-max_actions_per_traj: 20
+max_actions_per_traj: 10
 reward_normalization:
   grouping: tags # 可以tags(env_type)/traj_group_id(group)/batch(rollout_batch)... group_by计算reward/adv
   method: identity # asym_clip / identity / mean_std
@@ -173,15 +180,17 @@ custom_envs:
       is_slippery: false
 
 train_env_manager:
-  format_penalty: -0.001
-  env_groups: 64
+  format_penalty: 0
+  max_env_num_per_worker: 16
+  env_groups: 1024
   group_size: 1
   tags: [FrozenLake]
-  n_groups: [64] # If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
+  n_groups: [1024] # If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
 
 val_env_manager:
-  env_groups: 64
+  max_env_num_per_worker: 32
+  env_groups: 1024
   group_size: 1 # should be set to 1 because val temperature is set to 0 and same prompt leads to same output
   tags: [SimpleSokoban, LargerSokoban, SokobanDifferentGridVocab, FrozenLake]
-  n_groups: [16, 16, 16, 16] # TODO: If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
+  n_groups: [256, 256, 256, 256] # TODO: If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
