+++ examples/qwen2.5-0.5B-agentic_ds/agentic_val_sokoban.yaml
@@ -73,8 +73,15 @@ actor_train:
   data_args:
     template: qwen2_5
   strategy_args:
-    strategy_name: deepspeed_train
-    strategy_config: ${deepspeed_zero2}
+#    strategy_name: deepspeed_train
+#    strategy_config: ${deepspeed_zero3}
+    strategy_name: megatron_train
+    strategy_config:
+      tensor_model_parallel_size: 1
+      pipeline_model_parallel_size: 1
+      expert_model_parallel_size: 1
+      use_distributed_optimizer: true
+      recompute_granularity: full
   device_mapping: list(range(0,8))
   infer_batch_size: 2
 
@@ -84,7 +91,7 @@ actor_infer:
     disable_gradient_checkpointing: true
     dtype: bf16
   generating_args:
-    max_new_tokens: 32 # single-turn response length
+    max_new_tokens: 128 # single-turn response length
     top_p: 0.99
     top_k: 100
     num_beams: 1
@@ -118,7 +125,7 @@ enable_response_mask: True
 action_sep: "||"
 use_turn_scores: False # important to GAE when applying token-level rewards to token-level advantages. If False, will take the sum of scores as the reward for the last turn.
 enable_think: False # False -> no think RL
-max_actions_per_traj: 20
+max_actions_per_traj: 10
 reward_normalization:
   grouping: tags # 可以tags(env_type)/traj_group_id(group)/batch(rollout_batch)... group_by计算reward/adv
   method: identity # asym_clip / identity / mean_std
@@ -171,16 +178,18 @@ custom_envs:
       is_slippery: false
 
 train_env_manager:
-  format_penalty: -0.001
-  env_groups: 64
+  format_penalty: -0.15 # sokoban env penalty_for_step=-0.1
+  max_env_num_per_worker: 16
+  env_groups: 512
   # under the same group, the env config and env seed are ensured to be equal
   group_size: 1
   tags: [SimpleSokoban]
-  n_groups: [64] # If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
+  n_groups: [512] # If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
 
 val_env_manager:
-  env_groups: 64
+  max_env_num_per_worker: 32
+  env_groups: 1024
   group_size: 1 # should be set to 1 because val temperature is set to 0 and same prompt leads to same output
   tags: [SimpleSokoban, LargerSokoban, SokobanDifferentGridVocab, FrozenLake]
-  n_groups: [16, 16, 16, 16] # TODO: If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
+  n_groups: [256, 256, 256, 256] # TODO: If not set, all env names divide nums equally. Under the same group, the env config and env seed (prompt) are equal in each generation
